Split component and types
