+++ packages/react-native-gesture-handler/src/v3/components/index.ts
@@ -23,5 +23,9 @@ export {
 
 export { default as Pressable } from './Pressable';
 
-export { Clickable } from './Clickable';
-export type { ClickableProps } from './Clickable';
+export { Clickable } from './Clickable/Clickable';
+export {
+  ClickableAnimationMode,
+  ClickableOpacityMode,
+} from './Clickable/ClickableProps';
+export type { ClickableProps } from './Clickable/ClickableProps';
